requirements.txt
langchain
langgraph
chromadb
langchain-google-genai
langchain-community
langchain-tavily
tavily-python
faiss-cpu
python-dotenv
sentence-transformers
pypdf
langchain-text-splitters